scrapers/pyproject.toml
[tool.poetry]
name = "airline-scraper"
version = "0.1.0"
description = ""
authors = ["Alexandru Placinta <alexandru.placinta@nexthink.com>"]
readme = "README.md"
packages = [{include = "airline_scraper"}]

[tool.poetry.dependencies]
python = "^3.8"
scrapy = "2.7.1"
kafka-python = "2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
